Revised Student Controller
The student controller was revised to work as an independent controller not associated with a model.  The initial work is done but a lot of garbage cleanup is needed.  Also touched up some visual appeal and fixed a small bug where if a tag wasn't represented in a particular year, that year would collapse and all columns would slide to the left after it.